Moved individual implementations into their own files to make the code easier to maintain. Updated imports to use explicit relative imports for future compatiblity with Python 3.
