+++ storage/__init__.py
@@ -1,3 +1,7 @@
-from storage import get_storage  # noqa: F401
-from storage import register_storage_protocol  # noqa: F401
-from storage import register_swift_protocol  # noqa: F401
+from .storage import get_storage  # noqa: F401
+from .storage import register_storage_protocol  # noqa: F401
+from .swift_storage import register_swift_protocol  # noqa: F401
+from . import local_storage  # noqa: F401
+from . import swift_storage  # noqa: F401
+from . import ftp_storage  # noqa: F401
+from . import s3_storage  # noqa: F401
